+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Upgraded go to 1.17.6 and github.com/hashicorp/terraform-plugin-sdk/v2 to v2.10.1
 
+
+### Changed
+* Added Importer for b2_bucket and b2_application_key resource
+
 ## [0.7.1] - 2021-10-14
 
 ### Changed